+++ src/cli.ts
@@ -10,18 +10,20 @@ import search from './search';
 import tutorials from './tutorials';
 import publish from './publish';
 import checkForUpdate from './update';
+import validate from './validate';
 
 program
-  .version('0.6.0')
+  .version('0.7.0')
   .usage('[options] <keywords>')
   .option('-b, --build <path/to/tutorial.md>',
-    'tutorial markdown file', /^.+\.md$/i)
+  'tutorial markdown file', /^.+\.md$/i)
   .option('-c, --create <name>', 'tutorial name')
   .option('-p, --publish <version>',
-    'publish tutorial to npm with new version number')
+  'publish tutorial to npm with new version number')
   .option('-t, --tutorials', 'list of local tutorial packages')
   .option('-s, --search <query>', 'search for tutorial package')
   .option('-r, --run', 'run tutorial')
+  .option('-v, --validate', 'validate tutorial')
   .parse(process.argv);
 
 checkForUpdate();
@@ -68,6 +70,12 @@ if (program.build) {
   const version = program.args[0];
   publish(version);
 
+} else if (program.validate) {
+  if (!validate()) {
+    fail();
+  }
+
+  // help
 } else {
   program.help();
 }