+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.6.0] - WIP
+- allow for programmatic use
 
 ## [0.5.0] - 2016-05-02
 - bug fixes