+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.7.0] - WIP
+- validate packageJson function
+
 ## [0.6.0] - 2016-06-12
 - allow for programmatic use
 - working `> coderoad tutorials`
-- validate name function
 
 ## [0.5.0] - 2016-05-02
 - bug fixes